+++ ui/src/character_sheet/initiative.rs
@@ -3,7 +3,7 @@ use yew::{html, Component, ComponentLink, Html, Properties, ShouldRender};
 
 #[derive(Properties, Clone)]
 pub struct Initiative {
-    pub value: usize,
+    pub value: i64,
 }
 
 impl Component for Initiative {
